Histogram 2D, with find local max using its image...
We should not be using the histogram 2d image to get the local max, but rather the histogram itself.
We are using the image also in part because we are dealing with a rather visual algorithm: that dilates and erodes the image, finds the contours, and computes the center of mass of these (which are supposed to be the Max points in the histogram).