Including installation for metis in dockerfile
Seems like metis was a hidden dependency that doesn't break the installation. But gives error when trying to execute the tests. Adding metis to the dockerfile seems to have fixed everything. Not sure if some dependency of metis caused this or metis itself. For now this docker build seems to be stable.

The requirement for timezone still exists. Adding TZ env did not seem to solve the issue will raise another PR if the fix seems to work.